pyproject.toml
[project]
name = "todaysearthquakes"
version = "0.1.0"
description = "A Twitter bot that automatically posts a daily map of the previous day's earthquakes, providing real-time global seismic activity updates."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cartopy>=0.25.0",
    "geopandas>=1.1.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "tweepy>=4.16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[tool.ruff]
target-version = "py310"
line-length = 119

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",   # Checks for the absence of trailing commas. (conflict with the formatter)
    "D203",     # Checks for docstrings on class definitions that are not preceded by a blank line. (conflict with D211)
    "D212",     # Checks for docstring summary lines that are not positioned on the first physical line of the docstring. (conflict with D213)
]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"              # Like Black, use double quotes for strings.
indent-style = "space"              # Like Black, indent with spaces, rather than tabs.
skip-magic-trailing-comma = false   # Like Black, respect magic trailing commas.
line-ending = "auto"                # Like Black, automatically detect the appropriate line ending.
